requirements.txt
Django==2.1
django-extensions==2.1.0
django-rest-framework==0.1.0
djangorestframework==3.8.2
factory-boy==2.11.1
Faker==0.9.0
flake8==3.5.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
python-dateutil==2.7.3
pytz==2018.5
six==1.11.0
text-unidecode==1.2
